requirements.txt
black
discord
discord.py
web3
py-etherscan-api